User Story 1 Complete
Actually parsing numbers now!

Switched output to strings because its easier to concatenate strings and
there doesn’t seem to be a requirement to output a number.

Digit unit tests are now redundant and could go away, but since a Digit
is sort of its own object, I’m going to leave the coverage for it there.