Add fixture files for individual entries
Program should be able to open and parse these files into numbers.

I’m having a lot of trouble trying to figure out how to open local files
in Javascript. It seems like this is not built into Javascript natively
for security reasons, which makes sense.

Node’s File System module seems like the way to go, but getting it up
and running does’t seem to be as simple as `var fs = require(‘fs’);