refactor digits array population
Initial var digits = [1,2,3,4,5,6,7,8,9] declaration could have anything
in it since the values aren’t used. Using numbers here might be
dangerous since it would it be hard to figure out where there are coming
from, but it was the easiest way to get an array with the right number 
of elements. Perhaps an old school for(i=0; i<9, i++) loop would be
better.